fix(middleware): move implementation to middleware.ts and remove proxy.ts
